pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-servicenow"
version = "0.2.0"
description = "MCP server for ServiceNow - table, CMDB, system, and update set operations"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Josh Schuller"},
]
dependencies = [
    "mcp>=1.3.0",
    "requests>=2.28.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "anyio>=4.0.0",
]

[project.scripts]
mcp-servicenow = "servicenow_mcp.cli:main"

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-mock>=3.0", "ruff>=0.4.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/servicenow_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
